+++ SyrupSource/Syrup/Framework/Declarative/IBinder.cs
@@ -1,5 +1,5 @@
 ﻿namespace Syrup.Framework;
 
-namespace Syrup.Framework {
+namespace Syrup.Framework.Declarative {
     public interface IBinder { }
 }
