+++ SyrupSource/Syrup/Framework/ISyrupModule.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Syrup.Framework.Declarative;
 
 namespace Syrup.Framework {
     public interface ISyrupModule {
